+++ lib/World.spec.lua
@@ -40,6 +40,35 @@ end
 
 return function()
 	describe("World", function()
+		it("should be iterable", function()
+			local world = World.new()
+			local A = component()
+			local B = component()
+
+			local eA = world:spawn(A())
+			local eB = world:spawn(B())
+			local eAB = world:spawn(A(), B())
+
+			local count = 0
+			for id, data in world do
+				count += 1
+				if id == eA then
+					expect(data[A]).to.be.ok()
+					expect(data[B]).to.never.be.ok()
+				elseif id == eB then
+					expect(data[B]).to.be.ok()
+					expect(data[A]).to.never.be.ok()
+				elseif id == eAB then
+					expect(data[A]).to.be.ok()
+					expect(data[B]).to.be.ok()
+				else
+					error("unknown entity", id)
+				end
+			end
+
+			expect(count).to.equal(3)
+		end)
+
 		it("should have correct size", function()
 			local world = World.new()
 			world:spawn()
@@ -190,7 +219,7 @@ return function()
 			local expectedResults = {
 				nil,
 				{
-					0,
+					1,
 					{
 						new = {
 							generation = 1,
@@ -201,17 +230,17 @@ return function()
 					1,
 					{
 						new = {
+							generation = 2,
+						},
+						old = {
 							generation = 1,
 						},
 					},
 				},
 				{
-					0,
+					2,
 					{
 						new = {
-							generation = 2,
-						},
-						old = {
 							generation = 1,
 						},
 					},
@@ -221,15 +250,15 @@ return function()
 					1,
 					{
 						old = {
-							generation = 1,
+							generation = 2,
 						},
 					},
 				},
 				{
-					0,
+					2,
 					{
 						old = {
-							generation = 2,
+							generation = 1,
 						},
 					},
 				},
@@ -240,6 +269,7 @@ return function()
 			local additionalQuery = C
 			loop:scheduleSystem(function(w)
 				local ran = false
+
 				for entityId, record in w:queryChanged(A) do
 					if additionalQuery then
 						if w:get(entityId, additionalQuery) == nil then
@@ -270,7 +300,7 @@ return function()
 					local results = {}
 					for entityId, record in w:queryChanged(A) do
 						count += 1
-						results[entityId] = record
+						results[entityId - 1] = record
 					end
 
 					if count == 0 then
@@ -412,8 +442,11 @@ return function()
 			world:remove(two, Health)
 			world:despawn(one)
 
-			expect(snapshot[1][1]).to.equal(0)
-			expect(snapshot[2][1]).to.equal(2)
+			if snapshot[2][1] == 3 then
+				expect(snapshot[1][1]).to.equal(1)
+			else
+				expect(snapshot[2][1]).to.equal(1)
+			end
 		end)
 	end)
 end